requirements.txt
rocketmq-client
redis
